+++ src/core/canopus/index.ts
@@ -1,0 +1,7 @@
+export * from "./objective.js";
+export * from "./acceptance.js";
+export * from "./convergence.js";
+export * from "./runState.js";
+export * from "./engine.js";
+export type { ControlPlaneSpecDocument as CanopusSpecDocument } from "../controlPlane/specs.js";
+export { createDefaultControlPlaneDocument as createDefaultCanopusDocument, loadControlPlaneSpecDocument as loadCanopusSpecDocument, parseControlPlaneSpecDocument as parseCanopusSpecDocument, requireRunnableControlPlaneDocument as requireRunnableCanopusDocument, toCanopusPublicSpecDocument } from "../controlPlane/specs.js";